pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Starter template for making a new agent LangGraph."
authors = [
    { name = "Juan-Francisco Reyes", email = "pacoreyes@protonmail.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.0.0",
    "langgraph>=1.0.0",
    "neo4j>=6.1.0",
    "pinecone>=6.0.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.1",
    "curl-cffi",
    "chainlit>=2.9.5",
    "structlog>=25.1.0",
    "polars",
    "langgraph-cli[inmem]",
    "langgraph-api>=0.7.36",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["agent*"]

[tool.setuptools.package-dir]
"" = "src"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that require live external services (deselect with '-m \"not integration\"')",
]

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.4.7",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.2",
]
